+++ NEWS.md
@@ -1,3 +1,5 @@
+# repo.data (development version)
+
 # repo.data 0.1.5
 
 * Package will not generate errors if no internet connection is available.