+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Adds `cran_actions()` for CRAN's movements of packages.
 * Adds `cran_issues()` for CRAN's notifications of package maintainers.
+* Adds more tests to increase code coverage and reduce bugs.
 
 # repo.data 0.1.5